+++ packages/frontend/web/public/images/decorations/fairy-boat.png
