+++ packages/frontend/web/public/images/decorations/fairy-castle.png
