+++ packages/frontend/web/public/images/decorations/fairy-cloud.png
