+++ packages/frontend/web/public/images/decorations/fairy-hay.png
