+++ packages/frontend/web/public/images/decorations/fairy-house.png
