+++ packages/frontend/web/public/images/decorations/fairy-ice-cream-shop.png
