+++ packages/frontend/web/public/images/decorations/fairy-mushroom.png
