+++ packages/frontend/web/public/images/decorations/fairy-rock.png
