+++ packages/frontend/web/public/images/decorations/mythologic-boat.png
