+++ packages/frontend/web/public/images/decorations/mythologic-double-pillar.png
