+++ packages/frontend/web/public/images/decorations/mythologic-fountain.png
