+++ packages/frontend/web/public/images/decorations/mythologic-helmet.png
