+++ packages/frontend/web/public/images/decorations/mythologic-pillar.png
