+++ packages/frontend/web/public/images/decorations/mythologic-pond.png
