+++ packages/frontend/web/public/images/decorations/mythologic-potion.png
