+++ packages/frontend/web/public/images/decorations/mythologic-rock-2.png
