+++ packages/frontend/web/public/images/decorations/mythologic-rock.png
