+++ packages/frontend/web/public/images/decorations/mythologic-tree-2.png
