+++ packages/frontend/web/public/images/decorations/mythologic-tree-3.png
