+++ packages/frontend/web/public/images/decorations/mythologic-tree.png
