+++ packages/frontend/web/public/images/decorations/mythologic-vase.png
