+++ packages/frontend/web/public/images/decorations/shadow-candle-skull.png
