+++ packages/frontend/web/public/images/decorations/shadow-cathedral.png
