+++ packages/frontend/web/public/images/decorations/shadow-cave.png
