+++ packages/frontend/web/public/images/decorations/shadow-moon.png
