+++ packages/frontend/web/public/images/decorations/shadow-pond.png
