+++ packages/frontend/web/public/images/decorations/shadow-potion-2.png
