+++ packages/frontend/web/public/images/decorations/shadow-snake-skull.png
