+++ packages/frontend/web/public/images/decorations/shadow-spider.png
