+++ packages/frontend/web/public/images/decorations/shadow-tree.png
