+++ packages/frontend/web/public/images/decorations/winged-ash.png
