+++ packages/frontend/web/public/images/decorations/winged-cloud.png
