+++ packages/frontend/web/public/images/decorations/winged-halloween.png
