+++ packages/frontend/web/public/images/decorations/winged-magic-lamp.png
