+++ packages/frontend/web/public/images/decorations/winged-nest.png
