+++ packages/frontend/web/public/images/decorations/winged-palm.png
