+++ packages/frontend/web/public/images/decorations/winged-tower.png
