+++ packages/frontend/web/public/images/decorations/winged-volcano.png
