+++ packages/frontend/web/src/assets/images/fairy-zone/potion.png
