+++ packages/frontend/web/src/assets/images/fairy-zone/visitor-a.png
